+++ changelog.md
@@ -5,6 +5,7 @@
 - Removed `LuaManager` and moved everything from it into `LuaTrigger`
 - From above change, fixed bug with Stop triggers not working on the Lua trigger
 - Made a separate fix to stop the trigger when the player dies
+- Added `Dialog.show()` function
 
 ## v1.0.0-alpha.5